+++ src/main/java/com/codeup/blog/models/Post.java
@@ -1,7 +1,7 @@
 package com.codeup.blog.models;
 
 public class Post {
-    private Long id;
+    private int id;
     private String title;
     private String body;
 
@@ -10,7 +10,7 @@ public class Post {
         this.body = body;
     }
 
-    public Post(Long id, String title, String body) {
+    public Post(int id, String title, String body) {
         this.id = id;
         this.title = title;
         this.body = body;
@@ -18,11 +18,11 @@ public class Post {
 
     public Post(){}
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
